refactor(drivers): consolidate parseSemver/DBSemver into metadata.go
Moves the per-driver semver.go source and semver_internal_test.go unit
tests into each driver's existing metadata.go and internal_test.go,
across mysql, postgres, sqlserver, oracle, clickhouse, sqlite3, duckdb,
and rqlite. Pure relocation, no logic changes.
