+++ drivers/duckdb/internal_test.go
@@ -339,3 +339,28 @@ func TestConnParams(t *testing.T) {
 	require.ElementsMatch(t, []string{"true", "false"}, params["enable_external_access"])
 	require.ElementsMatch(t, []string{"true", "false"}, params["enable_object_cache"])
 }
+
+func TestParseSemver(t *testing.T) {
+	testCases := []struct {
+		raw     string
+		want    string
+		wantErr bool
+	}{
+		{raw: "v1.5.2", want: "v1.5.2"}, // DuckDB version() is v-prefixed
+		{raw: "1.1.3", want: "v1.1.3"},
+		{raw: "not-a-version", wantErr: true},
+		{raw: "", wantErr: true},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.raw, func(t *testing.T) {
+			got, err := parseSemver(tc.raw)
+			if tc.wantErr {
+				require.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			require.Equal(t, tc.want, got)
+		})
+	}
+}
