+++ drivers/duckdb/metadata.go
@@ -18,6 +18,7 @@ import (
 	"github.com/neilotoole/sq/libsq/driver"
 	"github.com/neilotoole/sq/libsq/source"
 	"github.com/neilotoole/sq/libsq/source/metadata"
+	"golang.org/x/mod/semver"
 )
 
 // SQL queries for DuckDB system catalog introspection.
@@ -1562,4 +1563,31 @@ SELECT view_name FROM duckdb_views() WHERE schema_name = current_schema() AND vi
 	}
 
 	return count > 0, nil
+}
+
+// semverRx matches a leading dotted-numeric version token (up to three parts).
+// The optional leading "v" covers DuckDB's v-prefixed version() output.
+var semverRx = regexp.MustCompile(`^v?(\d+(?:\.\d+){0,2})`)
+
+// parseSemver normalizes a DuckDB version() string to canonical semver (e.g.
+// "v1.5.2"). DuckDB's version() is already v-prefixed.
+func parseSemver(raw string) (string, error) {
+	m := semverRx.FindStringSubmatch(strings.TrimSpace(raw))
+	if m == nil {
+		return "", errz.Errorf("no semver in duckdb version string: %q", raw)
+	}
+	v := semver.Canonical("v" + m[1])
+	if !semver.IsValid(v) {
+		return "", errz.Errorf("invalid duckdb semver %q from %q", v, raw)
+	}
+	return v, nil
+}
+
+// DBSemver implements driver.SQLDriver.
+func (d *driveri) DBSemver(ctx context.Context, db sqlz.DB) (string, error) {
+	var raw string
+	if err := db.QueryRowContext(ctx, stmtVersion).Scan(&raw); err != nil {
+		return "", errw(err)
+	}
+	return parseSemver(raw)
 }
