--- drivers/duckdb/semver_internal_test.go
@@ -1,32 +1,0 @@
-package duckdb
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestParseSemver(t *testing.T) {
-	testCases := []struct {
-		raw     string
-		want    string
-		wantErr bool
-	}{
-		{raw: "v1.5.2", want: "v1.5.2"}, // DuckDB version() is v-prefixed
-		{raw: "1.1.3", want: "v1.1.3"},
-		{raw: "not-a-version", wantErr: true},
-		{raw: "", wantErr: true},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.raw, func(t *testing.T) {
-			got, err := parseSemver(tc.raw)
-			if tc.wantErr {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.Equal(t, tc.want, got)
-		})
-	}
-}
