+++ drivers/rqlite/internal_test.go
@@ -804,3 +804,28 @@ func Test_rewritePeerDiscoveryError(t *testing.T) {
 		})
 	}
 }
+
+func TestParseSemver(t *testing.T) {
+	testCases := []struct {
+		raw     string
+		want    string
+		wantErr bool
+	}{
+		{raw: "3.45.1", want: "v3.45.1"}, // rqlite reports its SQLite version
+		{raw: "3.46.0", want: "v3.46.0"},
+		{raw: "not-a-version", wantErr: true},
+		{raw: "", wantErr: true},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.raw, func(t *testing.T) {
+			got, err := parseSemver(tc.raw)
+			if tc.wantErr {
+				require.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			require.Equal(t, tc.want, got)
+		})
+	}
+}
