--- drivers/rqlite/semver.go
@@ -1,38 +1,0 @@
-package rqlite
-
-import (
-	"context"
-	"regexp"
-	"strings"
-
-	"golang.org/x/mod/semver"
-
-	"github.com/neilotoole/sq/libsq/core/errz"
-	"github.com/neilotoole/sq/libsq/core/sqlz"
-)
-
-// semverRx matches a leading dotted-numeric version token (up to three parts).
-var semverRx = regexp.MustCompile(`^v?(\d+(?:\.\d+){0,2})`)
-
-// parseSemver normalizes the SQLite version string rqlite reports to canonical
-// semver (e.g. "v3.45.1").
-func parseSemver(raw string) (string, error) {
-	m := semverRx.FindStringSubmatch(strings.TrimSpace(raw))
-	if m == nil {
-		return "", errz.Errorf("no semver in rqlite version string: %q", raw)
-	}
-	v := semver.Canonical("v" + m[1])
-	if !semver.IsValid(v) {
-		return "", errz.Errorf("invalid rqlite semver %q from %q", v, raw)
-	}
-	return v, nil
-}
-
-// DBSemver implements driver.SQLDriver.
-func (d *driveri) DBSemver(ctx context.Context, db sqlz.DB) (string, error) {
-	var raw string
-	if err := db.QueryRowContext(ctx, "SELECT sqlite_version()").Scan(&raw); err != nil {
-		return "", errw(err)
-	}
-	return parseSemver(raw)
-}
