+++ drivers/sqlite3/metadata.go
@@ -7,10 +7,12 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
+	"regexp"
 	"strings"
 
 	"github.com/neilotoole/sq/drivers/sqlite3/sqlparser"
 	"github.com/neilotoole/sq/libsq/core/debugz"
+	"github.com/neilotoole/sq/libsq/core/errz"
 	"github.com/neilotoole/sq/libsq/core/kind"
 	"github.com/neilotoole/sq/libsq/core/lg"
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
@@ -20,6 +22,7 @@ import (
 	"github.com/neilotoole/sq/libsq/core/stringz"
 	"github.com/neilotoole/sq/libsq/driver"
 	"github.com/neilotoole/sq/libsq/source/metadata"
+	"golang.org/x/mod/semver"
 )
 
 // recordMetaFromColumnTypes returns record.Meta for colTypes.
@@ -1113,4 +1116,30 @@ func getTypeOfColumn(ctx context.Context, db sqlz.DB, tblName, colName string) (
 	}
 
 	return colType, nil
+}
+
+// semverRx matches a leading dotted-numeric version token (up to three parts).
+var semverRx = regexp.MustCompile(`^v?(\d+(?:\.\d+){0,2})`)
+
+// parseSemver normalizes a SQLite version string to canonical semver (e.g.
+// "v3.45.1").
+func parseSemver(raw string) (string, error) {
+	m := semverRx.FindStringSubmatch(strings.TrimSpace(raw))
+	if m == nil {
+		return "", errz.Errorf("no semver in sqlite version string: %q", raw)
+	}
+	v := semver.Canonical("v" + m[1])
+	if !semver.IsValid(v) {
+		return "", errz.Errorf("invalid sqlite semver %q from %q", v, raw)
+	}
+	return v, nil
+}
+
+// DBSemver implements driver.SQLDriver.
+func (d *driveri) DBSemver(ctx context.Context, db sqlz.DB) (string, error) {
+	var raw string
+	if err := db.QueryRowContext(ctx, "SELECT sqlite_version()").Scan(&raw); err != nil {
+		return "", errw(err)
+	}
+	return parseSemver(raw)
 }
