--- drivers/sqlite3/semver_internal_test.go
@@ -1,32 +1,0 @@
-package sqlite3
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestParseSemver(t *testing.T) {
-	testCases := []struct {
-		raw     string
-		want    string
-		wantErr bool
-	}{
-		{raw: "3.45.1", want: "v3.45.1"},
-		{raw: "3.46.0", want: "v3.46.0"},
-		{raw: "not-a-version", wantErr: true},
-		{raw: "", wantErr: true},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.raw, func(t *testing.T) {
-			got, err := parseSemver(tc.raw)
-			if tc.wantErr {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.Equal(t, tc.want, got)
-		})
-	}
-}
