--- drivers/sqlserver/semver.go
@@ -1,39 +1,0 @@
-package sqlserver
-
-import (
-	"context"
-	"regexp"
-	"strings"
-
-	"golang.org/x/mod/semver"
-
-	"github.com/neilotoole/sq/libsq/core/errz"
-	"github.com/neilotoole/sq/libsq/core/sqlz"
-)
-
-// semverRx matches a leading dotted-numeric version token (up to three parts).
-var semverRx = regexp.MustCompile(`^v?(\d+(?:\.\d+){0,2})`)
-
-// parseSemver normalizes a SQL Server ProductVersion string to canonical semver
-// (e.g. "v16.0.4115"). ProductVersion is four-part (major.minor.build.revision);
-// the regex caps it at the first three parts.
-func parseSemver(raw string) (string, error) {
-	m := semverRx.FindStringSubmatch(strings.TrimSpace(raw))
-	if m == nil {
-		return "", errz.Errorf("no semver in sqlserver version string: %q", raw)
-	}
-	v := semver.Canonical("v" + m[1])
-	if !semver.IsValid(v) {
-		return "", errz.Errorf("invalid sqlserver semver %q from %q", v, raw)
-	}
-	return v, nil
-}
-
-// DBSemver implements driver.SQLDriver.
-func (d *driveri) DBSemver(ctx context.Context, db sqlz.DB) (string, error) {
-	var raw string
-	if err := db.QueryRowContext(ctx, "SELECT SERVERPROPERTY('ProductVersion')").Scan(&raw); err != nil {
-		return "", errw(err)
-	}
-	return parseSemver(raw)
-}
