--- drivers/sqlserver/semver_internal_test.go
@@ -1,32 +1,0 @@
-package sqlserver
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestParseSemver(t *testing.T) {
-	testCases := []struct {
-		raw     string
-		want    string
-		wantErr bool
-	}{
-		{raw: "16.0.4115.5", want: "v16.0.4115"}, // four-part; regex caps at three
-		{raw: "15.0.2000.5", want: "v15.0.2000"},
-		{raw: "not-a-version", wantErr: true},
-		{raw: "", wantErr: true},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.raw, func(t *testing.T) {
-			got, err := parseSemver(tc.raw)
-			if tc.wantErr {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.Equal(t, tc.want, got)
-		})
-	}
-}
